+++ composer.json
@@ -1,0 +1,28 @@
+{
+    "name": "nikolassv/pagination",
+    "version": "1.0.0",
+    "type": "library",
+    "description": "A PHP Library for creating paginations",
+    "keywords": ["pagination", "logarithmic pagination"],
+    "license": "LGPL-3.0",
+    "authors": [
+        {
+            "name": "Nikolas Schmidt-Voigt",
+            "email": "n.schmidtvoigt@googlemail.com",
+            "role": "Developer"
+        }
+    ],
+    "require": {
+        "php": ">=5.4.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagination\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Pagination\\Test\\": "tests/"
+        }
+    }
+}
