+++ src/PageItem.php
@@ -1,4 +1,9 @@
 <?php
+
+namespace Pagination;
+
+use \Pagination\PageItemInterface;
+
 /**************************************************************
  * Copyright notice
  *
@@ -23,27 +28,27 @@
 
 
 /**
- * the pagination factory interface
- *
- * pagination factories produce pagination iterator instances and fill them
- * with the appropriate values
+ * a simple page item class
  *
  * @author Nikolas Schmidt-Voigt <n.schmidtvoigt@googlemail.com>
  * @license LGPL-3.0 <http://opensource.org/licenses/LGPL-3.0>
  */
 
- interface PaginationFactoryInterface
- {
-	/**
-	 * produces a new pagination iterator instance
-	 *
-	 * @param	int	min	the lowest element in the pagination
-	 * @param	int max	the highest element in the pagination
-	 * @param	int steps	the number of items displayed in the pagination
-	 * @param	int current	the current page
-	 * @return	PaginationIteratorInterface	the new pagination iterator
-	 */
-	public static function makeNewPagination($max, $steps, $current, $min);
- }
+class PageItem implements PageItemInterface
+{
+	protected $pageNumber;
 
+	public function __construct($pageNumber)
+	{
+		if (!is_integer($pageNumber)) {
+			throw new InvalidValueException('page number must be an interger');
+		}
 
+		$this->pageNumber = $pageNumber;
+	}
+
+	public function getPageNumber()
+	{
+		return $this->pageNumber;
+	}
+}
