+++ src/PaginationIteratorInterface.php
@@ -1,4 +1,9 @@
 <?php
+
+namespace Pagination;
+
+use \Iterator;
+
 /**************************************************************
  * Copyright notice
  *
@@ -23,27 +28,40 @@
 
 
 /**
- * a simple page item class
+ * interface for pagination iterators
+ *
+ * pagination iterators will iterator over a set of items in a ascending
+ * list. for each item it will either return a page item or a gap item.
  *
  * @author Nikolas Schmidt-Voigt <n.schmidtvoigt@googlemail.com>
  * @license LGPL-3.0 <http://opensource.org/licenses/LGPL-3.0>
  */
 
-class PageItem implements PageItemInterface
+interface PaginationIteratorInterface extends Iterator
 {
-	protected $pageNumber;
+	/**
+	 * construct a new PaginationIterator object
+	 *
+	 * @param	array	elements	the elements which are shown in the pagination
+	 */
+	public function __construct(array $elements);
 
-	public function __construct($pageNumber)
-	{
-		if (!is_integer($pageNumber)) {
-			throw new InvalidValueException('page number must be an interger');
-		}
+	/**
+	 * sets the page item class
+	 *
+	 * items of the element set will be returned as instances of this class
+	 *
+	 * @param	string	class name of the page items
+	 */
+	public function setPageItemClass($pageItemClass);
 
-		$this->pageNumber = $pageNumber;
-	}
-
-	public function getPageNumber()
-	{
-		return $this->pageNumber;
-	}
+	/**
+	 * sets the gap item class
+	 *
+	 * gaps between the elements of the element set will be returned as instances
+	 * of this class
+	 *
+	 * @param	string	class name of the gap items
+	 */
+	public function setGapItemClass($gapItemClass);
 }
